python-http-talk/production-ready/data-service/requirements.txt
flask==2.0.1
werkzeug==2.0.1
requests==2.26.0
PyJWT==2.1.0
prometheus-flask-exporter==0.18.6
gunicorn==20.1.0
